refactor AssignTo, MoveModal and feature flag modal as prompt modals + add unit tests + remove store mechanisms that were dedicated to these modals
