+++ shell/components/__tests__/PromptModal.test.ts
@@ -22,6 +22,9 @@ import SloDialog from '@shell/dialog/SloDialog.vue';
 import WechatDialog from '@shell/dialog/WechatDialog.vue';
 import ImportDialog from '@shell/dialog/ImportDialog.vue';
 import SearchDialog from '@shell/dialog/SearchDialog.vue';
+import AssignToDialog from '@shell/dialog/AssignToDialog.vue';
+import FeatureFlagListDialog from '@shell/dialog/FeatureFlagListDialog.vue';
+import MoveNamespaceDialog from '@shell/dialog/MoveNamespaceDialog.vue';
 import { createStore } from 'vuex';
 
 jest.mock('@shell/utils/clipboard', () => {
@@ -96,6 +99,9 @@ describe('component: PromptModal', () => {
     ['WechatDialog', WechatDialog],
     ['ImportDialog', ImportDialog],
     ['SearchDialog', SearchDialog],
+    ['AssignToDialog', AssignToDialog],
+    ['FeatureFlagListDialog', FeatureFlagListDialog],
+    ['MoveNamespaceDialog', MoveNamespaceDialog],
   ])('prompt Modal should render modal %p', (modalName, component) => {
     document.body.innerHTML = '<div id="modals"></div>';
     const wrapper = mount(PromptModal,
