+++ shell/detail/namespace.vue
@@ -5,7 +5,6 @@ import ResourceTabs from '@shell/components/form/ResourceTabs';
 import { COUNT, WORKLOAD_TYPES } from '@shell/config/types';
 import { WORKLOAD_SCHEMA } from '@shell/config/schema';
 import { getStatesByType } from '@shell/plugins/dashboard-store/resource-class';
-import MoveModal from '@shell/components/MoveModal';
 import Tab from '@shell/components/Tabbed/Tab';
 import ResourceTable from '@shell/components/ResourceTable';
 import SortableTable from '@shell/components/SortableTable';
@@ -20,7 +19,6 @@ export default {
     ResourceTable,
     ResourceTabs,
     Loading,
-    MoveModal,
     Tab,
     SortableTable
   },
@@ -264,6 +262,5 @@ export default {
         />
       </Tab>
     </ResourceTabs>
-    <MoveModal />
   </div>
 </template>
